+++ MinecraftDemo/src/Core/Services/MovementSystem/MovementModes/FlyNoClipMovementMode.cpp
@@ -12,5 +12,5 @@ void FlyNoClipMovementMode::execute() {
 	GameObject* go = m_go;
 	MovementComponent* mc = m_mc;
 	mc->m_velocity = mc->m_targetVelocity;
-	go->m_position += (mc->m_velocity * (float)m_deltaTime);
+	go->m_position += (mc->m_velocity * m_deltaTime);
 }
